requirements.txt
streamlit
python-dotenv
pdfplumber
pdf2image
pytesseract
google-generativeai
Pillow
